Fix how themes are made available at startup
Thanks to Iris Garcia for informing me that the were not registered as
"known" before.  This was done in issue 2 on the GitHub mirror:
<https://github.com/protesilaos/ef-themes/issues/2>.

The code I am adding here needs to be reviewed to simplify all the
internal functions.  Thought it has been a very long day (again), so I
will do that tomorrow.