Add ef-themes-mixed-fonts user option
My original plan was to not implement such an option and instead direct
users to other packages such as 'mixed-pitch'.  However, doing this at
the theme level feels more natural and gives us greater control over
what actually needs to be rendered in 'fixed-pitch'.

Furthermore, adding such an option does not really make the themes
harder to maintain as it does not affect colour combinations.