Explicitly support 'bold' and 'italic' faces
The default definition of 'italic' has conditionality which we do not
want, such as drawing an underline if the font does not have italics.
Styling it here prevents that from happening: the font without italics
will simply use the upright style.

Adding 'bold' here is just for clarity.  Its default definition is fine.