Make 'org-indent' always monospaced
This is essential for 'org-indent-mode' to align entries correctly.

Thanks to Steve Molitor for reporting the issue 9 over at the GitHub
mirror: <https://github.com/protesilaos/ef-themes/issues/9>.  Steven
tested a sample file with 'modus-vivendi' and got the desired results.
It reminded me how I handle this face with the 'modus-themes'.