Add support for completion group titles
These are headings which are used by various commands, such as
'consult-buffer', 'consult-imenu', and 'consult-grep'.  As they denote a
group name, we apply the 'name' colour mapping.  And since they are
titles, we inherit from 'bold' to make them look like headings.

We cannot use the default style of combining the 'shadow' and 'italic'
faces as the result is not sufficiently distinct from the main
foreground value for the purposes of the completion interfaces (whereas
that design is okay with the modus-themes).

For theme-wide consistency, also check 'compilation-info', which styles
file names in Grep and related buffers.  Quote from commit ffb4cfb:

    Decouple 'compilation-info' from 'success'

    It made file names green in Grep and Xref buffers, which does not look
    good in any theme.  Using the "name" mapping gives us thematically
    consistent results.

    I inspected emacs.git to check if there is anything that uses
    'compilation-info' to convey colour-coding along the lines of the
    familiar red-green dichotomy, but could not find anything.  As such, I
    am treating this face in more flexible terms.

Thanks to Summer Emacs for the feedback which amounted to:

1. Treating group titles as headings (i.e. bold).
2. Using a more consistent colour.

[ Any mistakes are always my own. ]