Stop associating message-mml with "info" mapping
It would be rendered in what normally is a green colour, though this is
not appropriate here.  Instead, we need those elements to have a more
neutral colour, as they apply to different types of information.  The
bold weight ensures that the construct is not long in the already
colourful buffer.