Implement mail-5 mapping for email multiparts
This revises commits 1a38dac, d7c4613.  I did not like their combined
effect.  Using a bespoke colour gives me more control over the end
result.

We want the multiparts of emails to be noticeable, as they have
important information.  But we also want them to be different from the
other elements in the email's body.