Tweak rcirc-nick-in-message-full-line
Thanks to Philip Kaludercic for telling me about the downsides of using
the 'italic' face:
<https://lists.sr.ht/~protesilaos/ef-themes/%3C87edwphahl.fsf%40posteo.net%3E>.

I still have not managed to test if bold is enough, or if we need to
apply some extra colour.  I will do it as soon as possible (testing IRC
is hard because I normally don't use it).