Make 'ef-themes-toggle' useful out-of-the-box
The previous behaviour was not intuitive.  The problem was that if the
user had not yet set the 'ef-themes-to-toggle' user option, the command
would do nothing helpful.

By falling back to selecting a theme, we do something useful and also
provide a prompt that mentions the 'ef-themes-to-toggle'.

This is based on the discussion with Philip Kaludercic about revising
the 'ef-themes-toggle' command to have a more Do-What-I-Mean behaviour:
<https://lists.sr.ht/~protesilaos/ef-themes/%3C878rmwfs0e.fsf%40posteo.net%3E>.