Make rcirc server messages look like comments
We don't want to use 'shadow' as this is for elements that the user may
choose to dim even more (by modifying the face).  Whereas comments are
meant to be read.

Also, this is how rcirc-server does it by default, except it copies (and
hardcodes) the attributes of 'font-lock-comment-face' instead of
directly inheriting it.