ef-tritanopia-light: revise graph colours
The previous set was not optimal for blue-yellow colour deficiency.

These are better than before, but they will never be perfect in
practice because whatever mode displays graphs needs to be designed
from the outset with tritanopia in mind.

For example, the org-habit graph is BY DESIGN unsuitable for
blue-yellow colour deficiency because of the colour coding it
introduces combined with the way it displays its information.

We cannot fix that at the theme level.