Implement palette subset for colour-coded underlines
We cannot use the regular colours for errors, warning, notices because
those are optimised to apply to text.  Those used for text need to
render the glyphs legible, whereas for underlines we need to present a
more vivid colour.

These new values make it possible to cover flymake, flyspell and
related packages.  For most of the themes in the collection, this is
not needed as their default faces are "okay".  It is necessary,
though, for the deuteranopia and tritanopia themes so that they use
the appropriate colours.