Use the main foreground in org-agenda-current-time
The other mapping was giving us decent results, though this was a
matter of serendipity: there is a chance that a theme maps an
inappropriate colour to 'variable' (e.g. red, which will interfere
with deadlines).

This new design is just as legible, while it has the upside of
reducing colouration where it is not needed, letting other elements
get the attention they deserve.