Remove background from consult lines
The default value of 'consult-line-number-prefix' inherits from the
'line-number' face.  The Ef themes make the latter inherit from
'default' in order to have the lines increase/decrease in font size
when the user calls the 'text-scale-adjust' command.  This arrangement
meant that Consult was implicitly getting the main background which
caused commands like 'consult-line' to not be highlighted from their
absolute beginning but only after the line number.

Thanks to Daniel Mendler for bringing this matter to my attention:
<https://lists.sr.ht/~protesilaos/ef-themes/%3Cb03413a6-cb77-615d-145d-db4eb710bfca%40daniel-mendler.de%3E>.