Make eglot-diagnostic-tag-unnecessary-face look like a warning
By default it inherits the 'shadow' face, which makes it
counter-intuitive as it dims the text instead of bringing it to our
attention.  The intent of 'eglot-diagnostic-tag-unnecessary-face' is to
highlight unused symbols, so this is better presented as a warning.

Thanks to Augusto Stoffel for bringing this matter to my attention.
This was done via a private channel and the information is shared
with permission.