Make whitespace-mode considerably more subtle; add palette mappings
The previous style involved the use of a dim grey background.  While
this is good to spot invisible characters quickly, it is bad for users
who want to run 'whitespace-mode' at all times (e.g. for Python which
is space-sensitive).

We thus remove the backgrounds by default but provide the option to
reinstate them via palette overrides (as documented at length in the
manual).  To this end, we have new semantic colour mappings for
ordinary negative space and its invisible characters:

- bg-space
- fg-space
- bg-space-err