Remove backgrounds from Git commit style warnings and errors
These pertain to the commit messages in magit and vc-git that change
colour when the text exceeds a character limit. In Magit, the line
after the summary line (first line) must be empty, so text there gets
another colour.

The background was not needed to convey the message, given that the
warning styles I have picked all are distinct from the main text.