Use modus-themes-generate-palette in all themes
I did the changes with keyboard macros as usual. I now notice that
some themes have 26 changes while others have 28. This should probably
be due to some older inconsequential irregularities. I loaded each
theme right now and they all seem okay, so I will let this pass. If
anybody reports a bug, I am always happy to fix it.

At any rate, the main point for this change is to (i) ensure that we
get a complete palette while (ii) that we register the theme's palette
as a "core palette" because these are the correct semantics (e.g. that
ef-light ultimately may have inherited something from modus-operandi
is not relevant to the question "what is your core theme?").